MME inference pipeline
