+++ .vscode/settings.json
@@ -1,7 +1,7 @@
 {
     "python.analysis.extraPaths": [
-        "/afs/cs.wisc.edu/u/q/i/qinxinghao/project/VisionZip-exp/models/LLaVA",
-        "/afs/cs.wisc.edu/u/q/i/qinxinghao/project/VisionZip-exp/models/VisionZip"
+        "/home/w1nd519994824/project/VisionZip-exp/models/LLaVA",
+        "/home/w1nd519994824/project/VisionZip-exp/models/VisionZip"
     ],
-    "python.defaultInterpreterPath": "/afs/cs.wisc.edu/u/q/i/qinxinghao/miniconda3/envs/visionzip/bin/python" 
+    "python.defaultInterpreterPath": "/home/w1nd519994824/project/miniconda3/envs/visionzip/bin/python" 
 }
